Big changes
* more cleanup
* merged stochastic and deterministic models,
  switch behavior with `--deterministic` flag
* changed names of files and model classes

Results:
stochastic simpleconv gets to ~60% accuracy on cifar10
with `--lr .05`
