--- log/mnist/mnist_10class/layers.err
@@ -1,10 +1,0 @@
-usage: main.py [-h] --dataset DATASET --model MODEL [--epochs EPOCHS]
-               [--lr LR] [--momentum MOMENTUM] [--no-cuda] [--no-orthogonal]
-               [--seed SEED] [--log-interval LOG_INTERVAL] [--save-model]
-               [--save-location SAVE_LOCATION] [--num-samples NUM_SAMPLES]
-               [--num-labels NUM_LABELS] [--batch-size BATCH_SIZE]
-               [--test-batch-size TEST_BATCH_SIZE] [--input-size INPUT_SIZE]
-               [--t-passes T_PASSES] [--i-passes I_PASSES] [--plot-boundary]
-               --hidden-layers HIDDEN_LAYERS [HIDDEN_LAYERS ...]
-               [--set-classes SET_CLASSES [SET_CLASSES ...]]
-main.py: error: argument --hidden-layers: expected at least one argument
