--- log/mnist/pyramid_1pass.err
@@ -1,13 +1,0 @@
-nohup: ignoring input
-Traceback (most recent call last):
-  File "main.py", line 58, in <module>
-    main()
-  File "main.py", line 55, in main
-    run_model(model, args, criterion, train_loader, test_loader, num_labels, device, args.num_passes)
-  File "/home/bjk224/BSNN/run_model.py", line 58, in run_model
-    train(args, model, device, train_loader, optimizer, epoch, criterion, args.batch_size, num_passes)
-  File "/home/bjk224/BSNN/run_model.py", line 20, in train
-    loss = criterion(output, labels)
-  File "main.py", line 54, in criterion
-    return torch.sum(~torch.eq(x,y))
-RuntimeError: Expected object of backend CPU but got backend CUDA for argument #2 'other'
