--- log/mnist/pyramid_1pass.out
@@ -1,67 +1,0 @@
-Using device:  cuda
-Train Data Shape:  torch.Size([60000, 28, 28])
-Test Data Shape:  torch.Size([60000])
-Model:  BernoulliModel(
-  (layers): ListModule(
-    (0): BernoulliLayer(
-      (lin): Linear(in_features=784, out_features=100, bias=True)
-    )
-    (1): BernoulliLayer(
-      (lin): Linear(in_features=100, out_features=95, bias=True)
-    )
-    (2): BernoulliLayer(
-      (lin): Linear(in_features=95, out_features=90, bias=True)
-    )
-    (3): BernoulliLayer(
-      (lin): Linear(in_features=90, out_features=85, bias=True)
-    )
-    (4): BernoulliLayer(
-      (lin): Linear(in_features=85, out_features=80, bias=True)
-    )
-    (5): BernoulliLayer(
-      (lin): Linear(in_features=80, out_features=75, bias=True)
-    )
-    (6): BernoulliLayer(
-      (lin): Linear(in_features=75, out_features=70, bias=True)
-    )
-    (7): BernoulliLayer(
-      (lin): Linear(in_features=70, out_features=65, bias=True)
-    )
-    (8): BernoulliLayer(
-      (lin): Linear(in_features=65, out_features=60, bias=True)
-    )
-    (9): BernoulliLayer(
-      (lin): Linear(in_features=60, out_features=55, bias=True)
-    )
-    (10): BernoulliLayer(
-      (lin): Linear(in_features=55, out_features=50, bias=True)
-    )
-    (11): BernoulliLayer(
-      (lin): Linear(in_features=50, out_features=45, bias=True)
-    )
-    (12): BernoulliLayer(
-      (lin): Linear(in_features=45, out_features=40, bias=True)
-    )
-    (13): BernoulliLayer(
-      (lin): Linear(in_features=40, out_features=35, bias=True)
-    )
-    (14): BernoulliLayer(
-      (lin): Linear(in_features=35, out_features=30, bias=True)
-    )
-    (15): BernoulliLayer(
-      (lin): Linear(in_features=30, out_features=25, bias=True)
-    )
-    (16): BernoulliLayer(
-      (lin): Linear(in_features=25, out_features=20, bias=True)
-    )
-    (17): BernoulliLayer(
-      (lin): Linear(in_features=20, out_features=15, bias=True)
-    )
-    (18): BernoulliLayer(
-      (lin): Linear(in_features=15, out_features=10, bias=True)
-    )
-    (19): BernoulliLayer(
-      (lin): Linear(in_features=10, out_features=4, bias=True)
-    )
-  )
-)
